Adding abstract class to keep SQL functionality in PostgreSQL and MySQL
